+++ flumine/config.py
@@ -16,3 +16,5 @@ current_time = None  # used for backtesting
 raise_errors = False  # used for call_check_market / call_process_market_book
 
 max_execution_workers = 32  # max number of workers in execution thread pool
+
+async_place_orders = False  # async place orders